Add Landlock support
Landlock is a Linux security module designed to sandbox applications
such as GNU Tar.  It's goal is to help mitigate the impact of security
vulnerabilities, or just bugs: https://landlock.io

Landlock is now in linux-next and should then be part of Linux v5.13 .

Because Landlock is designed as a best-effort security (enabled when
available), nothing is changed if the build environment doesn't have
linux/landlock.h (provided by a future libc-dev package).  This also
means that there is no user-reachable option to enable or disable it:
Landlock is enabled if available in the running system.

This first step to sandbox GNU Tar is simple.  Restrictions are only
enforced, according to the user-supplied subcommand, once the archive is
opened.  Indeed, it would require more invasive changes to handle all
supplied file hierarchy arguments (e.g. files to add to an archive).
The value of this sandboxing is mainly to forbid a compromised GNU Tar
process (because of a malicious archive) to write or execute arbitrary
files.

Landlock only restricts file opening, so file descriptor that are opened
before the enforcement are still available.

This patch adds a new helper tar_dirfd() to open and get the file
descriptor used to extract an archive to.  This is required to mark this
file hierarchy as allowed for writing.

Test results from the GNU Tar test suite with this patch applied and a running
kernel with Landlock enabled: 200 tests were successful and 38 tests were
skipped.

Signed-off-by: Mickaël Salaün <mic@linux.microsoft.com>